Move `pluralize` method into "pure" helpers file (since we don't require `vscode`)
