+++ extensions/ql-vscode/src/helpers.ts
@@ -584,11 +584,3 @@ export async function* walkDirectory(dir: string): AsyncIterableIterator<string>
     }
   }
 }
-
-/**
- * Pluralizes a word.
- * Example: Returns "N repository" if N is one, "N repositories" otherwise.
- */
-export function pluralize(numItems: number | undefined, singular: string, plural: string): string {
-  return numItems !== undefined ? `${numItems} ${numItems === 1 ? singular : plural}` : '';
-}
