+++ extensions/ql-vscode/src/pure/helpers-pure.ts
@@ -55,3 +55,11 @@ export function getErrorStack(e: any) {
 export function asError(e: any): Error {
   return e instanceof Error ? e : new Error(String(e));
 }
+
+/**
+ * Pluralizes a word.
+ * Example: Returns "N repository" if N is one, "N repositories" otherwise.
+ */
+export function pluralize(numItems: number | undefined, singular: string, plural: string): string {
+  return numItems !== undefined ? `${numItems} ${numItems === 1 ? singular : plural}` : '';
+}
